refactor request matcher
