+++ pkg/filters/proxies/grpcproxy/requestmatch.go
@@ -18,56 +18,27 @@
 package grpcproxy
 
 import (
-	"fmt"
-	"hash/fnv"
-	"math/rand"
-	"regexp"
-	"strings"
-
+	"github.com/megaease/easegress/pkg/filters/proxies"
+	"github.com/megaease/easegress/pkg/protocols"
 	"github.com/megaease/easegress/pkg/protocols/grpcprot"
-
-	"github.com/megaease/easegress/pkg/logger"
 )
-
-// RequestMatcher is the interface to match requests.
-type RequestMatcher interface {
-	Match(req *grpcprot.Request) bool
-}
 
 // RequestMatcherSpec describe RequestMatcher
 type RequestMatcherSpec struct {
-	Policy          string                    `json:"policy" jsonschema:"omitempty,enum=,enum=general,enum=ipHash,enum=headerHash,enum=random"`
-	MatchAllHeaders bool                      `json:"matchAllHeaders" jsonschema:"omitempty"`
-	Headers         map[string]*StringMatcher `json:"headers" jsonschema:"omitempty"`
-	URLs            []*URLMatcher             `json:"urls" jsonschema:"omitempty"`
-	Permil          uint32                    `json:"permil" jsonschema:"omitempty,minimum=0,maximum=1000"`
-	HeaderHashKey   string                    `json:"headerHashKey" jsonschema:"omitempty"`
+	proxies.RequestMatcherBaseSpec `json:",inline"`
+	Methods                        []*proxies.StringMatcher `json:"methods" jsonschema:"omitempty"`
 }
 
 // Validate validtes the RequestMatcherSpec.
 func (s *RequestMatcherSpec) Validate() error {
-	if s.Policy == "general" || s.Policy == "" {
-		if len(s.Headers) == 0 {
-			return fmt.Errorf("headers is not specified")
-		}
-	} else if s.Permil == 0 {
-		return fmt.Errorf("permil is not specified")
+	if err := s.RequestMatcherBaseSpec.Validate(); err != nil {
+		return err
 	}
 
-	for _, v := range s.Headers {
-		if err := v.Validate(); err != nil {
-			return err
-		}
-	}
-
-	for _, r := range s.URLs {
+	for _, r := range s.Methods {
 		if err := r.Validate(); err != nil {
 			return err
 		}
-	}
-
-	if s.Policy == "headerHash" && s.HeaderHashKey == "" {
-		return fmt.Errorf("headerHash needs to specify headerHashKey")
 	}
 
 	return nil
@@ -80,90 +51,48 @@ func NewRequestMatcher(spec *RequestMatcherSpec) RequestMatcher {
 		matcher := &generalMatcher{
 			matchAllHeaders: spec.MatchAllHeaders,
 			headers:         spec.Headers,
-			urls:            spec.URLs,
+			methods:         spec.Methods,
 		}
 		matcher.init()
 		return matcher
-	case "ipHash":
-		return &ipHashMatcher{permill: spec.Permil}
-	case "headerHash":
-		return &headerHashMatcher{
-			permill:       spec.Permil,
-			headerHashKey: spec.HeaderHashKey,
-		}
-	case "random":
-		return &randomMatcher{permill: spec.Permil}
+	default:
+		return proxies.NewRequestMatcher(&spec.RequestMatcherBaseSpec)
 	}
-
-	logger.Errorf("BUG: unsupported probability policy: %s", spec.Policy)
-	return &ipHashMatcher{permill: spec.Permil}
-}
-
-// randomMatcher implements random request matcher.
-type randomMatcher struct {
-	permill uint32
-}
-
-// Match implements protocols.Matcher.
-func (rm randomMatcher) Match(req *grpcprot.Request) bool {
-	return rand.Uint32()%1000 < rm.permill
-}
-
-// headerHashMatcher implements header hash request matcher.
-type headerHashMatcher struct {
-	permill       uint32
-	headerHashKey string
-}
-
-// Match implements protocols.Matcher.
-func (hhm headerHashMatcher) Match(req *grpcprot.Request) bool {
-	v := req.RawHeader().GetFirst(hhm.headerHashKey)
-	hash := fnv.New32()
-	hash.Write([]byte(v))
-	return hash.Sum32()%1000 < hhm.permill
-}
-
-// ipHashMatcher implements IP address hash matcher.
-type ipHashMatcher struct {
-	permill uint32
-}
-
-// Match implements protocols.Matcher.
-func (iphm ipHashMatcher) Match(req *grpcprot.Request) bool {
-	ip := req.RealIP()
-	hash := fnv.New32()
-	hash.Write([]byte(ip))
-	return hash.Sum32()%1000 < iphm.permill
 }
 
 // generalMatcher implements general grpc matcher.
 type generalMatcher struct {
 	matchAllHeaders bool
-	headers         map[string]*StringMatcher
-	urls            []*URLMatcher
+	headers         map[string]*proxies.StringMatcher
+	methods         []*proxies.StringMatcher
 }
 
 func (gm *generalMatcher) init() {
 	for _, h := range gm.headers {
-		h.init()
+		h.Init()
 	}
 
-	for _, url := range gm.urls {
-		url.init()
+	for _, m := range gm.methods {
+		m.Init()
 	}
 }
 
 // Match implements protocols.Matcher.
-func (gm *generalMatcher) Match(req *grpcprot.Request) bool {
-	matched := false
-	if gm.matchAllHeaders {
-		matched = gm.matchAllHeader(req)
-	} else {
-		matched = gm.matchOneHeader(req)
+func (gm *generalMatcher) Match(req protocols.Request) bool {
+	grpcreq, ok := req.(*grpcprot.Request)
+	if !ok {
+		panic("not a grpc request")
 	}
 
-	if matched && len(gm.urls) > 0 {
-		matched = gm.matchURL(req)
+	matched := false
+	if gm.matchAllHeaders {
+		matched = gm.matchAllHeader(grpcreq)
+	} else {
+		matched = gm.matchOneHeader(grpcreq)
+	}
+
+	if matched && len(gm.methods) > 0 {
+		matched = gm.matchMethod(grpcreq)
 	}
 
 	return matched
@@ -209,98 +138,9 @@ func (gm *generalMatcher) matchAllHeader(req *grpcprot.Request) bool {
 	return true
 }
 
-func (gm *generalMatcher) matchURL(req *grpcprot.Request) bool {
-	for _, url := range gm.urls {
-		if url.Match(req) {
-			return true
-		}
-	}
-	return false
-}
-
-// URLMatcher defines the match rule of a grpc request
-type URLMatcher struct {
-	URL *StringMatcher `json:"url" jsonschema:"required"`
-}
-
-// Validate validates the MethodAndURLMatcher.
-func (r *URLMatcher) Validate() error {
-	return r.URL.Validate()
-}
-
-func (r *URLMatcher) init() {
-	r.URL.init()
-}
-
-// Match matches a request.
-func (r *URLMatcher) Match(req *grpcprot.Request) bool {
-	return r.URL.Match(req.FullMethod())
-}
-
-// StringMatcher defines the match rule of a string
-type StringMatcher struct {
-	Exact  string `json:"exact" jsonschema:"omitempty"`
-	Prefix string `json:"prefix" jsonschema:"omitempty"`
-	RegEx  string `json:"regex" jsonschema:"omitempty,format=regexp"`
-	Empty  bool   `json:"empty" jsonschema:"omitempty"`
-	re     *regexp.Regexp
-}
-
-// Validate validates the StringMatcher.
-func (sm *StringMatcher) Validate() error {
-	if sm.Empty {
-		if sm.Exact != "" || sm.Prefix != "" || sm.RegEx != "" {
-			return fmt.Errorf("empty is conflict with other patterns")
-		}
-		return nil
-	}
-
-	if sm.Exact != "" {
-		return nil
-	}
-
-	if sm.Prefix != "" {
-		return nil
-	}
-
-	if sm.RegEx != "" {
-		return nil
-	}
-
-	return fmt.Errorf("all patterns are empty")
-}
-
-func (sm *StringMatcher) init() {
-	if sm.RegEx != "" {
-		sm.re = regexp.MustCompile(sm.RegEx)
-	}
-}
-
-// Match matches a string.
-func (sm *StringMatcher) Match(value string) bool {
-	if sm.Empty && value == "" {
-		return true
-	}
-
-	if sm.Exact != "" && value == sm.Exact {
-		return true
-	}
-
-	if sm.Prefix != "" && strings.HasPrefix(value, sm.Prefix) {
-		return true
-	}
-
-	if sm.re == nil {
-		return false
-	}
-
-	return sm.re.MatchString(value)
-}
-
-// MatchAny return true if any of the values matches.
-func (sm *StringMatcher) MatchAny(values []string) bool {
-	for _, v := range values {
-		if sm.Match(v) {
+func (gm *generalMatcher) matchMethod(req *grpcprot.Request) bool {
+	for _, m := range gm.methods {
+		if m.Match(req.FullMethod()) {
 			return true
 		}
 	}
