+++ pkg/filters/proxies/grpcproxy/requestmatch_test.go
@@ -19,11 +19,9 @@ package grpcproxy
 
 import (
 	"context"
-	"fmt"
-	"math/rand"
-	"strconv"
 	"testing"
 
+	"github.com/megaease/easegress/pkg/filters/proxies"
 	"github.com/megaease/easegress/pkg/protocols/grpcprot"
 	"google.golang.org/grpc/metadata"
 
@@ -40,105 +38,25 @@ func TestRequestMatcherSpecValidate(t *testing.T) {
 	spec.Permil = 100
 	assert.Error(spec.Validate())
 
-	spec.Headers = map[string]*StringMatcher{}
-	spec.Headers["X-Test"] = &StringMatcher{
+	spec.Headers = map[string]*proxies.StringMatcher{}
+	spec.Headers["X-Test"] = &proxies.StringMatcher{
 		Empty: true,
 		Exact: "abc",
 	}
 	assert.Error(spec.Validate())
 
-	spec.Headers["X-Test"] = &StringMatcher{Exact: "abc"}
-	spec.URLs = append(spec.URLs, &URLMatcher{
-		URL: &StringMatcher{
-			Empty: true,
-			Exact: "abc",
-		},
+	spec.Headers["X-Test"] = &proxies.StringMatcher{Exact: "abc"}
+	spec.Methods = append(spec.Methods, &proxies.StringMatcher{
+		Empty: true,
+		Exact: "abc",
 	})
 	assert.Error(spec.Validate())
 
-	spec.URLs[0] = &URLMatcher{
-		URL: &StringMatcher{Empty: true},
-	}
+	spec.Methods[0] = &proxies.StringMatcher{Empty: true}
 	assert.Error(spec.Validate())
 
 	spec.HeaderHashKey = "X-Test"
 	assert.NoError(spec.Validate())
-}
-
-func TestRandomMatcher(t *testing.T) {
-	rand.Seed(0)
-
-	rm := NewRequestMatcher(&RequestMatcherSpec{
-		Policy: "random",
-		Permil: 100,
-	})
-
-	match := 0
-	for i := 0; i < 10000; i++ {
-		if rm.Match(nil) {
-			match++
-		}
-	}
-
-	if match < 900 || match > 1100 {
-		t.Error("random matcher is not working as configured")
-	}
-}
-
-func TestHeaderHashMatcher(t *testing.T) {
-	rm := NewRequestMatcher(&RequestMatcherSpec{
-		Policy:        "headerHash",
-		HeaderHashKey: "X-Test",
-		Permil:        100,
-	})
-
-	sm := grpcprot.NewFakeServerStream(metadata.NewIncomingContext(context.Background(), metadata.MD{}))
-	req := grpcprot.NewRequestWithServerStream(sm)
-
-	match := 0
-	for i := 0; i < 10000; i++ {
-		req.Header().Set("X-Test", strconv.Itoa(i))
-		if rm.Match(req) {
-			match++
-		}
-	}
-
-	if match < 900 || match > 1100 {
-		t.Error("header hash matcher is not working as configured")
-	}
-}
-
-func TestIPHashMatcher(t *testing.T) {
-	rm := NewRequestMatcher(&RequestMatcherSpec{
-		Policy: "unknownPolicy",
-		Permil: 100,
-	})
-	switch rm.(type) {
-	case *ipHashMatcher:
-		break
-	default:
-		t.Error("should create an ip hash matcher")
-	}
-
-	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Policy: "ipHash",
-		Permil: 100,
-	})
-
-	match := 0
-	for i := 0; i < 10000; i++ {
-		a, b := i/256, i%256
-		sm := grpcprot.NewFakeServerStream(metadata.NewIncomingContext(context.Background(), metadata.MD{}))
-		req := grpcprot.NewRequestWithServerStream(sm)
-		req.SetRealIP(fmt.Sprintf("192.168.%d.%d", a, b))
-		if rm.Match(req) {
-			match++
-		}
-	}
-
-	if match < 900 || match > 1100 {
-		t.Errorf("ip hash matcher is not working as configured")
-	}
 }
 
 func TestGeneralMatche(t *testing.T) {
@@ -146,10 +64,12 @@ func TestGeneralMatche(t *testing.T) {
 
 	// match all headers
 	rm := NewRequestMatcher(&RequestMatcherSpec{
-		MatchAllHeaders: true,
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			MatchAllHeaders: true,
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
+			},
 		},
 	})
 
@@ -163,10 +83,12 @@ func TestGeneralMatche(t *testing.T) {
 	assert.False(rm.Match(req))
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		MatchAllHeaders: true,
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Empty: true, Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			MatchAllHeaders: true,
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Empty: true, Exact: "test2"},
+			},
 		},
 	})
 
@@ -175,9 +97,11 @@ func TestGeneralMatche(t *testing.T) {
 
 	// match one header
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Empty: true, Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Empty: true, Exact: "test2"},
+			},
 		},
 	})
 	assert.True(rm.Match(req))
@@ -186,9 +110,11 @@ func TestGeneralMatche(t *testing.T) {
 	assert.True(rm.Match(req))
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
+			},
 		},
 	})
 	assert.False(rm.Match(req))
@@ -197,75 +123,28 @@ func TestGeneralMatche(t *testing.T) {
 	req.Header().Set("X-Test1", "test1")
 	req.SetFullMethod("/abc")
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
-		},
-		URLs: []*URLMatcher{
-			{
-				URL: &StringMatcher{
-					Exact: "/abc",
-				},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
 			},
+		},
+		Methods: []*proxies.StringMatcher{
+			{Exact: "/abc"},
 		},
 	})
 	assert.True(rm.Match(req))
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
-		},
-		URLs: []*URLMatcher{
-			{
-				URL: &StringMatcher{
-					Exact: "/abcd",
-				},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
 			},
+		},
+		Methods: []*proxies.StringMatcher{
+			{Exact: "/abcd"},
 		},
 	})
 	assert.False(rm.Match(req))
-}
-
-func TestStringMatcher(t *testing.T) {
-	assert := assert.New(t)
-
-	// validation
-	sm := &StringMatcher{Empty: true}
-	assert.NoError(sm.Validate())
-	sm.init()
-
-	sm = &StringMatcher{Empty: true, Exact: "abc"}
-	assert.Error(sm.Validate())
-
-	sm = &StringMatcher{}
-	assert.Error(sm.Validate())
-
-	sm = &StringMatcher{RegEx: "^abc[0-9]+$"}
-	assert.NoError(sm.Validate())
-	sm.init()
-
-	sm.Prefix = "/xyz"
-	assert.NoError(sm.Validate())
-
-	sm.Exact = "/abc"
-	assert.NoError(sm.Validate())
-
-	// match
-	sm = &StringMatcher{Empty: true}
-	assert.True(sm.Match(""))
-	assert.False(sm.Match("abc"))
-
-	sm = &StringMatcher{RegEx: "^abc[0-9]+$"}
-	sm.init()
-	assert.True(sm.Match("abc123"))
-	assert.False(sm.Match("abc123d"))
-
-	sm.Prefix = "/xyz"
-	assert.True(sm.Match("/xyz123"))
-	assert.False(sm.Match("/Xyz123"))
-
-	sm.Exact = "/hello"
-	assert.True(sm.Match("/hello"))
-	assert.False(sm.Match("/Hello"))
 }
