+++ pkg/filters/proxies/httpproxy/requestmatch.go
@@ -18,46 +18,22 @@
 package httpproxy
 
 import (
-	"fmt"
-	"hash/fnv"
-	"math/rand"
-	"regexp"
-	"strings"
-
-	"github.com/megaease/easegress/pkg/logger"
+	"github.com/megaease/easegress/pkg/filters/proxies"
+	"github.com/megaease/easegress/pkg/protocols"
 	"github.com/megaease/easegress/pkg/protocols/httpprot"
 	"github.com/megaease/easegress/pkg/util/stringtool"
 )
 
-// RequestMatcher is the interface to match requests.
-type RequestMatcher interface {
-	Match(req *httpprot.Request) bool
-}
-
 // RequestMatcherSpec describe RequestMatcher
 type RequestMatcherSpec struct {
-	Policy          string                    `json:"policy" jsonschema:"omitempty,enum=,enum=general,enum=ipHash,enum=headerHash,enum=random"`
-	MatchAllHeaders bool                      `json:"matchAllHeaders" jsonschema:"omitempty"`
-	Headers         map[string]*StringMatcher `json:"headers" jsonschema:"omitempty"`
-	URLs            []*MethodAndURLMatcher    `json:"urls" jsonschema:"omitempty"`
-	Permil          uint32                    `json:"permil" jsonschema:"omitempty,minimum=0,maximum=1000"`
-	HeaderHashKey   string                    `json:"headerHashKey" jsonschema:"omitempty"`
+	proxies.RequestMatcherBaseSpec `json:",inline"`
+	URLs                           []*MethodAndURLMatcher `json:"urls" jsonschema:"omitempty"`
 }
 
 // Validate validtes the RequestMatcherSpec.
 func (s *RequestMatcherSpec) Validate() error {
-	if s.Policy == "general" || s.Policy == "" {
-		if len(s.Headers) == 0 {
-			return fmt.Errorf("headers is not specified")
-		}
-	} else if s.Permil == 0 {
-		return fmt.Errorf("permil is not specified")
-	}
-
-	for _, v := range s.Headers {
-		if err := v.Validate(); err != nil {
-			return err
-		}
+	if err := s.RequestMatcherBaseSpec.Validate(); err != nil {
+		return err
 	}
 
 	for _, r := range s.URLs {
@@ -66,15 +42,11 @@ func (s *RequestMatcherSpec) Validate() error {
 		}
 	}
 
-	if s.Policy == "headerHash" && s.HeaderHashKey == "" {
-		return fmt.Errorf("headerHash needs to specify headerHashKey")
-	}
-
 	return nil
 }
 
 // NewRequestMatcher creates a new traffic matcher according to spec.
-func NewRequestMatcher(spec *RequestMatcherSpec) RequestMatcher {
+func NewRequestMatcher(spec *RequestMatcherSpec) proxies.RequestMatcher {
 	switch spec.Policy {
 	case "", "general":
 		matcher := &generalMatcher{
@@ -84,68 +56,21 @@ func NewRequestMatcher(spec *RequestMatcherSpec) RequestMatcher {
 		}
 		matcher.init()
 		return matcher
-	case "ipHash":
-		return &ipHashMatcher{permill: spec.Permil}
-	case "headerHash":
-		return &headerHashMatcher{
-			permill:       spec.Permil,
-			headerHashKey: spec.HeaderHashKey,
-		}
-	case "random":
-		return &randomMatcher{permill: spec.Permil}
+	default:
+		return proxies.NewRequestMatcher(&spec.RequestMatcherBaseSpec)
 	}
-
-	logger.Errorf("BUG: unsupported probability policy: %s", spec.Policy)
-	return &ipHashMatcher{permill: spec.Permil}
-}
-
-// randomMatcher implements random request matcher.
-type randomMatcher struct {
-	permill uint32
-}
-
-// Match implements protocols.Matcher.
-func (rm randomMatcher) Match(req *httpprot.Request) bool {
-	return rand.Uint32()%1000 < rm.permill
-}
-
-// headerHashMatcher implements header hash request matcher.
-type headerHashMatcher struct {
-	permill       uint32
-	headerHashKey string
-}
-
-// Match implements protocols.Matcher.
-func (hhm headerHashMatcher) Match(req *httpprot.Request) bool {
-	v := req.HTTPHeader().Get(hhm.headerHashKey)
-	hash := fnv.New32()
-	hash.Write([]byte(v))
-	return hash.Sum32()%1000 < hhm.permill
-}
-
-// ipHashMatcher implements IP address hash matcher.
-type ipHashMatcher struct {
-	permill uint32
-}
-
-// Match implements protocols.Matcher.
-func (iphm ipHashMatcher) Match(req *httpprot.Request) bool {
-	ip := req.RealIP()
-	hash := fnv.New32()
-	hash.Write([]byte(ip))
-	return hash.Sum32()%1000 < iphm.permill
 }
 
 // generalMatcher implements general HTTP matcher.
 type generalMatcher struct {
 	matchAllHeaders bool
-	headers         map[string]*StringMatcher
+	headers         map[string]*proxies.StringMatcher
 	urls            []*MethodAndURLMatcher
 }
 
 func (gm *generalMatcher) init() {
 	for _, h := range gm.headers {
-		h.init()
+		h.Init()
 	}
 
 	for _, url := range gm.urls {
@@ -154,16 +79,21 @@ func (gm *generalMatcher) init() {
 }
 
 // Match implements protocols.Matcher.
-func (gm *generalMatcher) Match(req *httpprot.Request) bool {
+func (gm *generalMatcher) Match(req protocols.Request) bool {
+	httpreq, ok := req.(*httpprot.Request)
+	if !ok {
+		panic("BUG: not a http request")
+	}
+
 	matched := false
 	if gm.matchAllHeaders {
-		matched = gm.matchAllHeader(req)
+		matched = gm.matchAllHeader(httpreq)
 	} else {
-		matched = gm.matchOneHeader(req)
+		matched = gm.matchOneHeader(httpreq)
 	}
 
 	if matched && len(gm.urls) > 0 {
-		matched = gm.matchURL(req)
+		matched = gm.matchURL(httpreq)
 	}
 
 	return matched
@@ -220,8 +150,8 @@ func (gm *generalMatcher) matchURL(req *httpprot.Request) bool {
 
 // MethodAndURLMatcher defines the match rule of a http request
 type MethodAndURLMatcher struct {
-	Methods []string       `json:"methods" jsonschema:"omitempty,uniqueItems=true,format=httpmethod-array"`
-	URL     *StringMatcher `json:"url" jsonschema:"required"`
+	Methods []string               `json:"methods" jsonschema:"omitempty,uniqueItems=true,format=httpmethod-array"`
+	URL     *proxies.StringMatcher `json:"url" jsonschema:"required"`
 }
 
 // Validate validates the MethodAndURLMatcher.
@@ -230,7 +160,7 @@ func (r *MethodAndURLMatcher) Validate() error {
 }
 
 func (r *MethodAndURLMatcher) init() {
-	r.URL.init()
+	r.URL.Init()
 }
 
 // Match matches a request.
@@ -242,74 +172,4 @@ func (r *MethodAndURLMatcher) Match(req *httpprot.Request) bool {
 	}
 
 	return r.URL.Match(req.URL().Path)
-}
-
-// StringMatcher defines the match rule of a string
-type StringMatcher struct {
-	Exact  string `json:"exact" jsonschema:"omitempty"`
-	Prefix string `json:"prefix" jsonschema:"omitempty"`
-	RegEx  string `json:"regex" jsonschema:"omitempty,format=regexp"`
-	Empty  bool   `json:"empty" jsonschema:"omitempty"`
-	re     *regexp.Regexp
-}
-
-// Validate validates the StringMatcher.
-func (sm *StringMatcher) Validate() error {
-	if sm.Empty {
-		if sm.Exact != "" || sm.Prefix != "" || sm.RegEx != "" {
-			return fmt.Errorf("empty is conflict with other patterns")
-		}
-		return nil
-	}
-
-	if sm.Exact != "" {
-		return nil
-	}
-
-	if sm.Prefix != "" {
-		return nil
-	}
-
-	if sm.RegEx != "" {
-		return nil
-	}
-
-	return fmt.Errorf("all patterns are empty")
-}
-
-func (sm *StringMatcher) init() {
-	if sm.RegEx != "" {
-		sm.re = regexp.MustCompile(sm.RegEx)
-	}
-}
-
-// Match matches a string.
-func (sm *StringMatcher) Match(value string) bool {
-	if sm.Empty && value == "" {
-		return true
-	}
-
-	if sm.Exact != "" && value == sm.Exact {
-		return true
-	}
-
-	if sm.Prefix != "" && strings.HasPrefix(value, sm.Prefix) {
-		return true
-	}
-
-	if sm.re == nil {
-		return false
-	}
-
-	return sm.re.MatchString(value)
-}
-
-// MatchAny return true if any of the values matches.
-func (sm *StringMatcher) MatchAny(values []string) bool {
-	for _, v := range values {
-		if sm.Match(v) {
-			return true
-		}
-	}
-	return false
 }
