+++ pkg/filters/proxies/httpproxy/requestmatch_test.go
@@ -18,12 +18,10 @@
 package httpproxy
 
 import (
-	"fmt"
-	"math/rand"
 	"net/http"
-	"strconv"
 	"testing"
 
+	"github.com/megaease/easegress/pkg/filters/proxies"
 	"github.com/megaease/easegress/pkg/protocols/httpprot"
 	"github.com/stretchr/testify/assert"
 )
@@ -38,16 +36,16 @@ func TestRequestMatcherSpecValidate(t *testing.T) {
 	spec.Permil = 100
 	assert.Error(spec.Validate())
 
-	spec.Headers = map[string]*StringMatcher{}
-	spec.Headers["X-Test"] = &StringMatcher{
+	spec.Headers = map[string]*proxies.StringMatcher{}
+	spec.Headers["X-Test"] = &proxies.StringMatcher{
 		Empty: true,
 		Exact: "abc",
 	}
 	assert.Error(spec.Validate())
 
-	spec.Headers["X-Test"] = &StringMatcher{Exact: "abc"}
+	spec.Headers["X-Test"] = &proxies.StringMatcher{Exact: "abc"}
 	spec.URLs = append(spec.URLs, &MethodAndURLMatcher{
-		URL: &StringMatcher{
+		URL: &proxies.StringMatcher{
 			Empty: true,
 			Exact: "abc",
 		},
@@ -55,7 +53,7 @@ func TestRequestMatcherSpecValidate(t *testing.T) {
 	assert.Error(spec.Validate())
 
 	spec.URLs[0] = &MethodAndURLMatcher{
-		URL: &StringMatcher{Empty: true},
+		URL: &proxies.StringMatcher{Empty: true},
 	}
 	assert.Error(spec.Validate())
 
@@ -63,92 +61,17 @@ func TestRequestMatcherSpecValidate(t *testing.T) {
 	assert.NoError(spec.Validate())
 }
 
-func TestRandomMatcher(t *testing.T) {
-	rand.Seed(0)
-
-	rm := NewRequestMatcher(&RequestMatcherSpec{
-		Policy: "random",
-		Permil: 100,
-	})
-
-	match := 0
-	for i := 0; i < 10000; i++ {
-		if rm.Match(nil) {
-			match++
-		}
-	}
-
-	if match < 900 || match > 1100 {
-		t.Error("random matcher is not working as configured")
-	}
-}
-
-func TestHeaderHashMatcher(t *testing.T) {
-	rm := NewRequestMatcher(&RequestMatcherSpec{
-		Policy:        "headerHash",
-		HeaderHashKey: "X-Test",
-		Permil:        100,
-	})
-
-	stdr, _ := http.NewRequest(http.MethodGet, "http://megaease.com/abc", nil)
-	req, _ := httpprot.NewRequest(stdr)
-
-	match := 0
-	for i := 0; i < 10000; i++ {
-		stdr.Header.Set("X-Test", strconv.Itoa(i))
-		if rm.Match(req) {
-			match++
-		}
-	}
-
-	if match < 900 || match > 1100 {
-		t.Error("header hash matcher is not working as configured")
-	}
-}
-
-func TestIPHashMatcher(t *testing.T) {
-	rm := NewRequestMatcher(&RequestMatcherSpec{
-		Policy: "unknownPolicy",
-		Permil: 100,
-	})
-	switch rm.(type) {
-	case *ipHashMatcher:
-		break
-	default:
-		t.Error("should create an ip hash matcher")
-	}
-
-	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Policy: "ipHash",
-		Permil: 100,
-	})
-
-	stdr := &http.Request{Header: http.Header{}}
-
-	match := 0
-	for i := 0; i < 10000; i++ {
-		a, b := i/256, i%256
-		stdr.Header.Set("X-Real-Ip", fmt.Sprintf("192.168.%d.%d", a, b))
-		req, _ := httpprot.NewRequest(stdr)
-		if rm.Match(req) {
-			match++
-		}
-	}
-
-	if match < 900 || match > 1100 {
-		t.Errorf("ip hash matcher is not working as configured")
-	}
-}
-
 func TestGeneralMatche(t *testing.T) {
 	assert := assert.New(t)
 
 	// match all headers
 	rm := NewRequestMatcher(&RequestMatcherSpec{
-		MatchAllHeaders: true,
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			MatchAllHeaders: true,
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
+			},
 		},
 	})
 
@@ -162,10 +85,12 @@ func TestGeneralMatche(t *testing.T) {
 	assert.False(rm.Match(req))
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		MatchAllHeaders: true,
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Empty: true, Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			MatchAllHeaders: true,
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Empty: true, Exact: "test2"},
+			},
 		},
 	})
 
@@ -174,9 +99,11 @@ func TestGeneralMatche(t *testing.T) {
 
 	// match one header
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Empty: true, Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Empty: true, Exact: "test2"},
+			},
 		},
 	})
 	assert.True(rm.Match(req))
@@ -185,9 +112,11 @@ func TestGeneralMatche(t *testing.T) {
 	assert.True(rm.Match(req))
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
+			},
 		},
 	})
 	assert.False(rm.Match(req))
@@ -196,14 +125,16 @@ func TestGeneralMatche(t *testing.T) {
 	stdr.Header.Set("X-Test1", "test1")
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
+			},
 		},
 		URLs: []*MethodAndURLMatcher{
 			{
 				Methods: []string{http.MethodGet},
-				URL: &StringMatcher{
+				URL: &proxies.StringMatcher{
 					Exact: "/abc",
 				},
 			},
@@ -212,14 +143,16 @@ func TestGeneralMatche(t *testing.T) {
 	assert.True(rm.Match(req))
 
 	rm = NewRequestMatcher(&RequestMatcherSpec{
-		Headers: map[string]*StringMatcher{
-			"X-Test1": {Exact: "test1"},
-			"X-Test2": {Exact: "test2"},
+		RequestMatcherBaseSpec: proxies.RequestMatcherBaseSpec{
+			Headers: map[string]*proxies.StringMatcher{
+				"X-Test1": {Exact: "test1"},
+				"X-Test2": {Exact: "test2"},
+			},
 		},
 		URLs: []*MethodAndURLMatcher{
 			{
 				Methods: []string{http.MethodGet},
-				URL: &StringMatcher{
+				URL: &proxies.StringMatcher{
 					Exact: "/abcd",
 				},
 			},
@@ -232,7 +165,7 @@ func TestMethodAndURLMatcher(t *testing.T) {
 	assert := assert.New(t)
 
 	m := &MethodAndURLMatcher{
-		URL: &StringMatcher{
+		URL: &proxies.StringMatcher{
 			Exact: "/abc",
 		},
 	}
@@ -250,47 +183,4 @@ func TestMethodAndURLMatcher(t *testing.T) {
 
 	m.Methods = []string{http.MethodPost}
 	assert.False(m.Match(req))
-}
-
-func TestStringMatcher(t *testing.T) {
-	assert := assert.New(t)
-
-	// validation
-	sm := &StringMatcher{Empty: true}
-	assert.NoError(sm.Validate())
-	sm.init()
-
-	sm = &StringMatcher{Empty: true, Exact: "abc"}
-	assert.Error(sm.Validate())
-
-	sm = &StringMatcher{}
-	assert.Error(sm.Validate())
-
-	sm = &StringMatcher{RegEx: "^abc[0-9]+$"}
-	assert.NoError(sm.Validate())
-	sm.init()
-
-	sm.Prefix = "/xyz"
-	assert.NoError(sm.Validate())
-
-	sm.Exact = "/abc"
-	assert.NoError(sm.Validate())
-
-	// match
-	sm = &StringMatcher{Empty: true}
-	assert.True(sm.Match(""))
-	assert.False(sm.Match("abc"))
-
-	sm = &StringMatcher{RegEx: "^abc[0-9]+$"}
-	sm.init()
-	assert.True(sm.Match("abc123"))
-	assert.False(sm.Match("abc123d"))
-
-	sm.Prefix = "/xyz"
-	assert.True(sm.Match("/xyz123"))
-	assert.False(sm.Match("/Xyz123"))
-
-	sm.Exact = "/hello"
-	assert.True(sm.Match("/hello"))
-	assert.False(sm.Match("/Hello"))
 }
