+++ pkg/object/meshcontroller/spec/spec.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/megaease/easegress/pkg/cluster/customdata"
 	"github.com/megaease/easegress/pkg/filters/mock"
+	"github.com/megaease/easegress/pkg/filters/proxies"
 	proxy "github.com/megaease/easegress/pkg/filters/proxies/httpproxy"
 	"github.com/megaease/easegress/pkg/filters/ratelimiter"
 	"github.com/megaease/easegress/pkg/resilience"
@@ -229,9 +230,9 @@ type (
 
 	// CanaryRule is one matching rule for canary.
 	CanaryRule struct {
-		ServiceInstanceLabels map[string]string               `json:"serviceInstanceLabels" jsonschema:"required"`
-		Headers               map[string]*proxy.StringMatcher `json:"headers" jsonschema:"required"`
-		URLs                  []*urlrule.URLRule              `json:"urls" jsonschema:"required"`
+		ServiceInstanceLabels map[string]string                 `json:"serviceInstanceLabels" jsonschema:"required"`
+		Headers               map[string]*proxies.StringMatcher `json:"headers" jsonschema:"required"`
+		URLs                  []*urlrule.URLRule                `json:"urls" jsonschema:"required"`
 	}
 
 	// ServiceCanary is the service canary entry.
@@ -247,7 +248,7 @@ type (
 
 	// TrafficRules is the rules of traffic.
 	TrafficRules struct {
-		Headers map[string]*proxy.StringMatcher `json:"headers" jsonschema:"required"`
+		Headers map[string]*proxies.StringMatcher `json:"headers" jsonschema:"required"`
 	}
 
 	// LoadBalance is the spec of service load balance.
@@ -480,7 +481,7 @@ func (sc ServiceCanary) Validate() error {
 
 // Clone clones TrafficRules.
 func (tr *TrafficRules) Clone() *TrafficRules {
-	headers := map[string]*proxy.StringMatcher{}
+	headers := map[string]*proxies.StringMatcher{}
 	for k, v := range tr.Headers {
 		stringMatch := *v
 		headers[k] = &stringMatch
